Removed FAQ test
